move client assertion test helpers into client_assertion_credentials::test
